+++ packages/ppx_ptree/test/cases/constructor_attribute.ml
@@ -1,0 +1,1 @@
+type t = A | C of int * string [@ptree.int] [@@deriving ptree]